requirements.txt
Appium-Python-Client==2.1.2
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
greenlet==1.1.2
h11==0.13.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
outcome==1.1.0
pycparser==2.21
pyOpenSSL==22.0.0
selenium==4.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.32
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
Werkzeug==2.0.3
wsproto==1.0.0
